requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
detecto==1.2.2
DetectoInference==0.0.1
ecdsa==0.19.2
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyfiglet==1.0.4
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
starlette==1.0.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uvicorn==0.42.0